Extracted an interface.
